Changed `app_interface` to `mediator`
